+++ src/components/Sancionados/Index.js
@@ -7,7 +7,7 @@ import Grid from "@material-ui/core/Grid/Grid";
 import img1 from "../../assets/img/servidores_publicos_sancionados.svg";
 import img2 from "../../assets/img/particulares_sancionados.svg";
 import img3 from "../../assets/img/servidores_visualizaciones.svg";
-import TablaServidoresSancionados from './Servidores/TablaServidoresSancionados';
+import TablaServidoresSancionados from './Servidores/BuscadorServidoresSancionados';
 import TablaParticularesSancionados from './Particulares/TablaParticularesSancionados';
 import Header from './Header/Header';
 import Dashboard from "./dashboard/Servidores/Dashboard";
