+++ src/components/Sancionados/Servidores/BusquedaServidor.js
@@ -20,6 +20,14 @@ import Checkbox from '@material-ui/core/Checkbox';
 import Input from '@material-ui/core/Input';
 import ExpandMoreIcon from '@material-ui/icons/ExpandMore';
 import ExpandLessIcon from '@material-ui/icons/ExpandLess';
+import rp from "request-promise";
+import Modal from "@material-ui/core/Modal";
+import CircularProgress from "@material-ui/core/CircularProgress";
+import MensajeErrorDatos from "../../Tablas/MensajeErrorDatos";
+import Switch from "@material-ui/core/Switch";
+import Collapse from "@material-ui/core/Collapse";
+import Previos from "../../Tablas/Previos";
+import TablaServidoresSancionados from "./TablaServidoresSancionados";
 
 const styles = theme => ({
     formControl: {
@@ -33,6 +41,27 @@ const styles = theme => ({
         justifyContent: 'center',
         alignItems: 'center'
     },
+    progress: {
+        position: 'fixed',
+        margin: 'auto',
+        left: 0,
+        right: 0,
+        top: 0,
+        bottom: 0
+    },
+    desc: {
+        color: theme.palette.primary.dark,
+    },
+    container: {
+        marginTop: '30px',
+        marginBottom: '30px',
+        overflowX: 'auto',
+    },
+    section: {
+        maxWidth: '1200px',
+        overflowX: 'auto',
+        padding: theme.spacing(1)
+    },
 });
 
 const tiposSancion = [
@@ -42,7 +71,7 @@ const tiposSancion = [
     {label: 'Destitución del empleo, cargo o comisión', value: 'D'}
 ]
 
-const camposOrdenamiento=[
+const camposOrdenamiento = [
     {label: 'RFC', value: 'rfc'},
     {label: 'CURP', value: 'curp'},
     {label: 'Nombre', value: 'nombres'},
@@ -50,6 +79,7 @@ const camposOrdenamiento=[
     {label: 'Apellido Dos', value: 'segundoApellido'},
     {label: 'Institución', value: 'institucionDependencia'}
 ]
+
 const tiposOrdenamiento = [
     {label: 'Ascendente', value: 'ASC'},
     {label: 'Descendente', value: 'DESC'}
@@ -58,19 +88,221 @@ const tiposOrdenamiento = [
 class BusquedaServidor extends React.Component {
     constructor(props) {
         super(props);
-        this.state={
-            busquedaAvanzada : false
+        this.state = {
+            filterData: null,
+            page: 0,
+            rowsPerPage: 10,
+            totalRows: 0,
+            panelPrevios: true,
+            error: false,
+            loading: false,
+            busquedaAvanzada: false,
+            //Filtros
+            nombresServidor: '',
+            apellidoUno: '',
+            apellidoDos: '',
+            tipoSancion: [],
+            rfc: '',
+            curp: '',
+            institucionDependencia: "",
+            nivel: 'todos',
+            campoOrden: '',
+            tipoOrden: '',
+            institucionesLista: []
         }
     }
 
-    handleBusquedaAvanzada=()=>{
-        this.setState({busquedaAvanzada: !this.state.busquedaAvanzada })
+    componentDidMount() {
+        this.loadInstituciones();
     }
 
+    loadInstituciones = () => {
+        let sug = [];
+        let options = {
+            uri: process.env.REACT_APP_HOST_PDNBACK + '/apis/getDependenciasServidores',
+            json: true,
+            method: "post",
+            body: {
+                nivel: this.state.nivel
+            }
+        };
+        rp(options)
+            .then(data => {
+                data.data.forEach((item, index) => {
+                    sug.push({value: item, label: item, key: index});
+                });
+                this.setState({institucionesLista: sug, institucionDependencia: ''});
+            }).catch(err => {
+            this.setState({error: true})
+        });
+    }
+
+    handleChangeCampo = (varState, event) => {
+        this.setState({
+            [varState]: event ? (event.target ? event.target.value : event.value) : ''
+        }, () => {
+            switch (varState) {
+                case 'nivel':
+                    this.loadInstituciones();
+                    break;
+                case 'campoOrden':
+                    this.setState({tipoOrden: event.target.value ? {label: 'Ascendente', value: 'ASC'} : ''});
+                    break;
+                case 'tipoOrden':
+                    this.setState({campoOrden: event.target.value ? {label: 'RFC', value: 'rfc'} : ''});
+                    break;
+            }
+        })
+    };
+
+    handleCleanAll = () => {
+        this.setState(
+            {
+                filterData: null,
+                previos: null,
+                nivel: 'todos',
+                tipoSancion: [],
+                nombresServidor: '',
+                institucionDependencia: '',
+                apellidoUno: '',
+                apellidoDos: '',
+                rfc: '',
+                curp: '',
+                campoOrden: '',
+                tipoOrden: ''
+            }, () => {
+                this.loadInstituciones();
+            })
+    };
+
+    handleSearchPrevios = () => {
+        this.setState({
+            loading: true,
+            filterData: null,
+            //previos: null,
+        }, () => {
+            let body =
+                {
+                    "filtros": this.makeFiltros(),
+                    "limit": 1,
+                    "offset": 0,
+                    "nivel": this.state.nivel
+                };
+
+            let options = {
+                method: 'POST',
+                uri: process.env.REACT_APP_HOST_PDNBACK + '/apis/getPrevioServidoresSancionados',
+                json: true,
+                body: body
+            };
+            rp(options)
+                .then(res => {
+                    this.setState({previos: res, loading: false, error: false})
+                }).catch(err => {
+                 this.setState({loading: false, error: true});
+            });
+        });
+    };
+
+    makeFiltros = () => {
+        let filtros = {};
+        let {institucionDependencia, nombresServidor, primerApellido, segundoApellido, rfc, curp, tipoSancion} = this.state;
+        if (nombresServidor) filtros.nombres = nombresServidor;
+        if (primerApellido) filtros.primer_apellido = primerApellido;
+        if (segundoApellido) filtros.segundo_apellido = segundoApellido;
+        if (rfc) filtros.rfc = rfc;
+        if (curp) filtros.curp = curp;
+        if (institucionDependencia && institucionDependencia !== '') filtros.nombre = institucionDependencia;
+        if (tipoSancion.length > 0) filtros.tipoSancion = tipoSancion.map(item => item.value);
+        return filtros;
+    };
+
+    makeSort = () => {
+        let sort = {};
+        if (this.state.campoOrden && this.state.tipoOrden) sort[this.state.campoOrden.value] = this.state.tipoOrden.value;
+        return sort;
+    };
+
+    handleChange = () => {
+        this.setState({
+            panelPrevios: !this.state.panelPrevios
+        })
+    }
+
+    handleError = (val) => {
+        this.setState({
+            error: val
+        })
+    }
+
+    handleBusquedaAvanzada = () => {
+        this.setState({busquedaAvanzada: !this.state.busquedaAvanzada})
+    }
+
+    handleSearchAPI = () => {
+        this.setState({loading: true}, () => {
+            let body =
+                {
+                    "filtros": this.makeFiltros(),
+                    "limit": this.state.rowsPerPage,
+                    "offset": this.state.rowsPerPage * this.state.page,
+                    "clave_api": this.state.api,
+                    "sort": this.makeSort()
+                };
+
+            let options = {
+                method: 'POST',
+                uri: process.env.REACT_APP_HOST_PDNBACK + '/apis/getServidoresSancionados',
+                json: true,
+                body: body
+            };
+
+            rp(options)
+                .then(res => {
+                    this.setState({
+                        filterData: res.data,
+                        loading: false,
+                        totalRows: res.totalRows,
+                        error: false
+                    },)
+                }).catch(err => {
+                this.setState({loading: false, error: true});
+            });
+        });
+
+    };
+
+    handleChangeAPI = (val) => {
+        this.setState({
+            api: val,
+            page: 0
+        }, () => {
+            this.handleSearchAPI()
+        });
+    };
+
+    handleChangePage = (event, page) => {
+        this.setState({page}, () => {
+            this.handleSearchAPI();
+        });
+    };
+
+    handleChangeRowsPerPage = event => {
+        this.setState({rowsPerPage: event.target.value, page: 0}, () => {
+            this.handleSearchAPI();
+        });
+    };
+    verDetalle = (event, elemento) => {
+        this.setState({elementoSeleccionado: elemento, open: true});
+    };
+
+
     render() {
-        const {classes, handleChangeCampo, nombreServidor, apellidoUno, apellidoDos, rfc, curp, institucion, nivel, tipoSancion,campoOrden,tipoOrden, institucionesLista} = this.props;
+        const {classes} = this.props;
+        const {nombresServidor, apellidoUno, apellidoDos, rfc, curp, institucionDependencia, nivel, tipoSancion, campoOrden, tipoOrden, institucionesLista} = this.state;
         return (
             <div>
+                {/*Buscador*/}
                 <Grid container spacing={4}>
                     <Grid item xs={12}>
                         <Typography><b>Busca un servidor público sancionado</b></Typography>
@@ -81,8 +313,8 @@ class BusquedaServidor extends React.Component {
                                 id="search"
                                 label="Nombre(s)"
                                 type="search"
-                                onChange={(e) => handleChangeCampo('nombresServidor', e)}
-                                value={nombreServidor}
+                                onChange={(e) => this.handleChangeCampo('nombresServidor', e)}
+                                value={nombresServidor}
                             />
 
                         </FormControl>
@@ -93,7 +325,7 @@ class BusquedaServidor extends React.Component {
                                 id="search"
                                 label="Apellido Uno"
                                 type="search"
-                                onChange={(e) => handleChangeCampo('primerApellido', e)}
+                                onChange={(e) => this.handleChangeCampo('apellidoUno', e)}
                                 value={apellidoUno}
                             />
 
@@ -105,7 +337,7 @@ class BusquedaServidor extends React.Component {
                                 id="search"
                                 label="Apellido Dos"
                                 type="search"
-                                onChange={(e) => handleChangeCampo('segundoApellido', e)}
+                                onChange={(e) => this.handleChangeCampo('apellidoDos', e)}
                                 value={apellidoDos}
                             />
                         </FormControl>
@@ -118,7 +350,7 @@ class BusquedaServidor extends React.Component {
                                     id="demo-mutiple-checkbox"
                                     multiple
                                     value={tipoSancion}
-                                    onChange={e => handleChangeCampo('tipoSancion', e)}
+                                    onChange={e => this.handleChangeCampo('tipoSancion', e)}
                                     input={<Input/>}
                                     renderValue={
                                         selected => {
@@ -126,7 +358,6 @@ class BusquedaServidor extends React.Component {
                                                 return <em>Cualquiera</em>;
                                             }
                                             return selected.map(element => element.label).join(', ')
-                                            //selected.join(', ')
                                         }
                                     }
 
@@ -150,8 +381,8 @@ class BusquedaServidor extends React.Component {
                             <InputLabel shrink id="demo-simple-select-placeholder-label-label">
                                 Institución
                             </InputLabel>
-                            <Select value={institucion}
-                                    onChange={(e) => handleChangeCampo('institucionDependencia', e)}
+                            <Select value={institucionDependencia}
+                                    onChange={(e) => this.handleChangeCampo('institucionDependencia', e)}
                                     displayEmpty
                             >
                                 <MenuItem value="" key={-1}><em>Cualquiera</em></MenuItem>
@@ -174,7 +405,7 @@ class BusquedaServidor extends React.Component {
                                         name="gender1"
                                         className={classes.group}
                                         value={nivel}
-                                        onChange={(e) => handleChangeCampo('nivel', e)}
+                                        onChange={(e) => this.handleChangeCampo('nivel', e)}
                             >
                                 <FormControlLabel value="todos" control={<Radio/>} label="Todos"/>
                                 <FormControlLabel value="federal" control={<Radio/>} label="Federal"/>
@@ -185,7 +416,7 @@ class BusquedaServidor extends React.Component {
                     </Grid>
                     <Grid item xs={12}>
                         <Button onClick={() => this.handleBusquedaAvanzada()}
-                                startIcon={this.state.busquedaAvanzada?<ExpandLessIcon/>:<ExpandMoreIcon />}
+                                startIcon={this.state.busquedaAvanzada ? <ExpandLessIcon/> : <ExpandMoreIcon/>}
                         >Búsqueda avanzada</Button>
                     </Grid>
 
@@ -195,7 +426,7 @@ class BusquedaServidor extends React.Component {
                                 id="search"
                                 label="CURP"
                                 type="search"
-                                onChange={(e) => handleChangeCampo('curp', e)}
+                                onChange={(e) => this.handleChangeCampo('curp', e)}
                                 value={curp}
                             />
 
@@ -207,19 +438,19 @@ class BusquedaServidor extends React.Component {
                                 id="search"
                                 label="RFC"
                                 type="search"
-                                onChange={(e) => handleChangeCampo('rfc', e)}
+                                onChange={(e) => this.handleChangeCampo('rfc', e)}
                                 value={rfc}
                             />
 
                         </FormControl>
                     </Grid>}
-                    {this.state.busquedaAvanzada  && <Grid item xs={12} md={3}>
+                    {this.state.busquedaAvanzada && <Grid item xs={12} md={3}>
                         <FormControl className={classes.formControl}>
                             <InputLabel shrink id="demo-mutiple-checkbox-label">Ordenar por</InputLabel>
                             <Select displayEmpty
                                     id="demo-mutiple-checkbox"
                                     value={campoOrden}
-                                    onChange={e => handleChangeCampo('campoOrden', e)}
+                                    onChange={e => this.handleChangeCampo('campoOrden', e)}
                                     input={<Input/>}
                                     renderValue={
                                         selected => {
@@ -231,7 +462,7 @@ class BusquedaServidor extends React.Component {
                                     }
 
                             >
-                                <MenuItem  value={''}>
+                                <MenuItem value={''}>
                                     <em>Ninguno</em>
                                 </MenuItem>
                                 {camposOrdenamiento.map(tipo => (
@@ -249,7 +480,7 @@ class BusquedaServidor extends React.Component {
                             <Select displayEmpty
                                     id="demo-mutiple-checkbox"
                                     value={tipoOrden}
-                                    onChange={e => handleChangeCampo('tipoOrden', e)}
+                                    onChange={e => this.handleChangeCampo('tipoOrden', e)}
                                     input={<Input/>}
                                     renderValue={
                                         selected => {
@@ -261,7 +492,7 @@ class BusquedaServidor extends React.Component {
                                     }
 
                             >
-                                <MenuItem  value={''}>
+                                <MenuItem value={''}>
                                     <em>Ninguno</em>
                                 </MenuItem>
                                 {tiposOrdenamiento.map(tipo => (
@@ -276,18 +507,71 @@ class BusquedaServidor extends React.Component {
 
                     <Grid item md={10}/>
                     <Grid item xs={12} md={1} className={classes.centrado}>
-                        <Button variant="contained" color="secondary" className={classes.button} onClick={()=>this.props.handleSearch('FIELD_FILTER')}>
+                        <Button variant="contained" color="secondary" className={classes.button}
+                                onClick={() => this.handleSearchPrevios()}>
                             Buscar
                         </Button>
                     </Grid>
                     <Grid item xs={12} md={1} className={classes.centrado}>
                         <Button variant="contained" color="secondary" className={classes.button}
-                                onClick={()=>this.props.handleCleanAll()}>
+                                onClick={() => this.handleCleanAll()}>
                             Limpiar
                         </Button>
                     </Grid>
+                    <Grid item xs={12}>
+                        {
+                            this.state.loading &&
+                            <Modal
+                                open={this.state.loading}
+                                disableAutoFocus={true}
+                            >
+                                <CircularProgress className={classes.progress} id="spinnerLoading" size={200}/>
+                            </Modal>
 
+                        }
+                        {
+                            this.state.error && <MensajeErrorDatos/>
+                        }
+                    </Grid>
                 </Grid>
+                {/*PREVIOS*/}
+                {this.state.previos && this.state.previos.length > 0 &&
+                <Grid container>
+                    <Grid item xs={12} className={classes.section}>
+                            <FormControlLabel
+                                control={<Switch className={classes.containerPrevios}
+                                                 checked={this.state.panelPrevios}
+                                                 onChange={() => this.handleChange()}/>}
+                                label={
+                                    <Typography variant="h6" className={classes.desc}>
+                                        {this.state.panelPrevios ? 'Ocultar resultados generales' : 'Mostrar resultados generales'}</Typography>}
+                            />
+                    </Grid>
+                    <Grid item xs={12} className={classes.section}>
+                            <div className={classes.container}>
+                                <Collapse in={this.state.panelPrevios}>
+                                    <Previos previos={this.state.previos} handleChangeAPI={this.handleChangeAPI}/>
+                                </Collapse>
+                            </div>
+                    </Grid>
+                </Grid>
+                }
+                {/*TABLA*/}
+                {this.state.filterData && this.state.filterData.length > 0 &&
+                <Grid container>
+                    <Grid item xs={12} className={classes.section}>
+                        <div className={classes.container} id={'test'}>
+                            <TablaServidoresSancionados data={this.state.filterData} page={this.state.page}
+                                                        rowsPerPage={this.state.rowsPerPage}
+                                                        totalRows={this.state.totalRows}
+                                                        handleChangePage={this.handleChangePage}
+                                                        handleChangeRowsPerPage={this.handleChangeRowsPerPage}
+                                                        verDetalle={this.props.verDetalle}/>
+                        </div>
+
+                    </Grid>
+                </Grid>
+                }
             </div>
         );
     }
